fix(menu): add en-US indicator for curriculum
For the other links with icon, it is less trivial, because
those are currently using an inline-flex with a gap, which
would cause the indicator to be vertically centered, and
have an unexpected gap next to them.

We also cannot add the indicator behind Blog, because there is
not necessarily enough space, so this might break the layout.